bot/requirements.txt
aiogram==3.18.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.0
